fix(table): Fixes an issue with partial keyframes on expandable row.
This was quite impossible to reproduce reliably and to be honest, we are not entirely sure if this will fix the problem.
Regarding to this issue in the angular components, this should help though and we were not seeing the error after this fix in our development environment.
https://github.com/angular/components/issues/13970

Fixes APM-276923